refactor: Split createTable.js into two files
